+++ libcloudvagrant/compute/driver.py
@@ -369,11 +369,10 @@ class VagrantDriver(base.NodeDriver):
         """
         self.log.info("Destroying node '%s' ..", node.name)
         try:
+            for v in self.list_volumes():
+                if v.attached_to == node.name:
+                    self.detach_volume(v)
             with self._catalogue as c:
-                for v in c.get_volumes():
-                    if v.attached_to == node.name:
-                        self.log.debug("destroy_node(): Detaching %s", v)
-                        self.detach_volume(v)
                 self._vagrant("destroy --force", node.name)
                 for ip in node._public_ips + node._private_ips:
                     self.log.debug("destroy_node(): Deallocating address %s",